粗略完成Product的controller部分，修改了service方法类型不正确的小错误

+++ productservice/src/main/java/com/seckill/productservice/service/IFinancialProductService.java
@@ -13,13 +13,13 @@ public interface IFinancialProductService{
     boolean addFinancialProduct(FinancialProductEntity financialProductEntity);
 
     //删除理财产品（根据ID）
-    boolean deleteFinancialProduct(int financialProductId);
+    boolean deleteFinancialProduct(long financialProductId);
 
     //更新理财产品
     boolean updateFinancialProduct(FinancialProductEntity financialProductEntity);
 
     //查找理财产品（根据ID）
-    FinancialProductEntity findFinancialProductById(int financialProductId);
+    FinancialProductEntity findFinancialProductById(long financialProductId);
 
     //查找所有理财产品
     List<FinancialProductEntity> findAll();
